moving folder data build to folder.js
