+++ nx/blocks/media-library/utils/filters.js
@@ -52,6 +52,9 @@ export function getAvailableFilters() {
 function createSearchSuggestion(item) {
   if (!item.name && !item.url && !item.doc) return null;
 
+  // Exclude SVG files from search suggestions (consistent with 'all' filter)
+  if (isSvgFile(item)) return null;
+
   return {
     type: 'media',
     value: item,
@@ -63,132 +66,6 @@ function createSearchSuggestion(item) {
       type: getMediaType(item),
     },
   };
-}
-
-/**
- * Build folder hierarchy from document path
- * @param {Map} hierarchy - Hierarchy map to populate
- * @param {string} docPath - Document path
- */
-function buildFolderHierarchy(hierarchy, docPath) {
-  if (!docPath) return;
-
-  // Remove leading slash
-  const cleanPath = docPath.startsWith('/') ? docPath.substring(1) : docPath;
-  const parts = cleanPath.split('/').filter(Boolean);
-
-  if (parts.length === 0) return;
-
-  // Simple rule: if the last part ends with .html, it's a file
-  // Everything else in the path are folders
-  const lastPart = parts[parts.length - 1];
-  const isFile = lastPart.endsWith('.html');
-
-  if (isFile) {
-    // Create folders for all parts except the last one
-    const folderParts = parts.slice(0, -1);
-
-    let currentPath = '';
-    folderParts.forEach((part) => {
-      currentPath = currentPath ? `${currentPath}/${part}` : part;
-
-      if (!hierarchy.has(currentPath)) {
-        hierarchy.set(currentPath, {
-          path: currentPath,
-          name: part,
-          level: currentPath.split('/').length,
-          children: new Set(),
-          parent: currentPath.includes('/') ? currentPath.substring(0, currentPath.lastIndexOf('/')) : null,
-          count: 0,
-          type: 'folder',
-          hasFiles: false,
-        });
-      }
-    });
-
-    // Add the file itself to the hierarchy
-    const filePath = cleanPath;
-    if (!hierarchy.has(filePath)) {
-      hierarchy.set(filePath, {
-        path: filePath,
-        name: lastPart,
-        level: parts.length,
-        children: new Set(),
-        parent: folderParts.length > 0 ? folderParts.join('/') : null,
-        count: 0,
-        type: 'file',
-        hasFiles: false,
-      });
-    }
-
-    // Always update the parent-child relationship, even if file already exists
-    if (folderParts.length > 0) {
-      const parentPath = folderParts.join('/');
-      if (hierarchy.has(parentPath)) {
-        hierarchy.get(parentPath).hasFiles = true;
-        hierarchy.get(parentPath).children.add(filePath);
-      }
-    }
-  }
-
-  // Build parent-child relationships for both folders and files
-  hierarchy.forEach((node, path) => {
-    if (node.parent && hierarchy.has(node.parent)) {
-      hierarchy.get(node.parent).children.add(path);
-    }
-  });
-}
-
-/**
- * Calculate media counts for each folder in the hierarchy
- * @param {Map} hierarchy - Folder hierarchy map
- * @param {Array} mediaData - Media data array
- */
-function calculateFolderCounts(hierarchy, mediaData) {
-  if (!hierarchy || !mediaData) return;
-
-  // Reset all counts
-  hierarchy.forEach((folder) => {
-    folder.count = 0;
-  });
-
-  // Count media items for each folder and file
-  mediaData.forEach((media) => {
-    if (media.doc) {
-      const docPath = media.doc;
-
-      // Remove leading slash
-      const cleanPath = docPath.startsWith('/') ? docPath.substring(1) : docPath;
-      const parts = cleanPath.split('/').filter(Boolean);
-
-      if (parts.length === 0) return;
-
-      // If it's a file (ends with .html), count for the file itself and all parent folders
-      const lastPart = parts[parts.length - 1];
-      if (lastPart.endsWith('.html')) {
-        const filePath = cleanPath;
-        const folderParts = parts.slice(0, -1); // All parts except the file
-
-        // Count for the file itself
-        const file = hierarchy.get(filePath);
-        if (file) {
-          file.count += 1;
-        }
-
-        // Count for all parent folders
-        let currentPath = '';
-        folderParts.forEach((part) => {
-          currentPath = currentPath ? `${currentPath}/${part}` : part;
-          const folder = hierarchy.get(currentPath);
-          if (folder) {
-            folder.count += 1;
-          }
-        });
-      }
-    }
-  });
-
-  // Counts are now calculated for both folders and files
 }
 
 // ============================================================================
@@ -207,7 +84,6 @@ export function processMediaData(mediaData) {
       searchSuggestions: [],
       usageMap: new Map(),
       folderHierarchy: new Map(),
-      docPaths: new Set(),
       mediaTypes: new Set(),
     };
   }
@@ -216,8 +92,6 @@ export function processMediaData(mediaData) {
   const filterCounts = {};
   const searchSuggestions = [];
   const usageMap = new Map();
-  const folderHierarchy = new Map();
-  const docPaths = new Set();
   const mediaTypes = new Set();
 
   // Initialize filter counts
@@ -258,10 +132,9 @@ export function processMediaData(mediaData) {
       }
     }
 
-    // 4. Build folder hierarchy
+    // 4. Collect document paths (for folder dialog to use later)
     if (item.doc) {
       docPaths.add(item.doc);
-      buildFolderHierarchy(folderHierarchy, item.doc);
     }
 
     // 5. Collect media types
@@ -271,8 +144,7 @@ export function processMediaData(mediaData) {
     }
   });
 
-  // 6. Calculate folder counts after building hierarchy
-  calculateFolderCounts(folderHierarchy, mediaData);
+  // Note: Folder hierarchy and counts moved to folder dialog component for on-demand processing
 
   // Sort search suggestions by relevance
   searchSuggestions.sort((a, b) => {
@@ -290,7 +162,7 @@ export function processMediaData(mediaData) {
     filterCounts,
     searchSuggestions: searchSuggestions.slice(0, 50), // Limit suggestions
     usageMap,
-    folderHierarchy,
+    folderHierarchy: new Map(), // Empty map - folder dialog will build its own
     docPaths: Array.from(docPaths).sort(),
     mediaTypes: Array.from(mediaTypes),
   };
